fix: refactoring tests
